Add multikey spec & remove sign verify spec.
